generate version using git describe
- use `git describe` to determine the semver-compatible version.

This helps to ensure that we can easily distinguish the version shown
for a released build from versions shown for other builds beyond a
release tag. The earlier practice of bumping the version in a separate
commit after each release was not strong enough. This also allows for
greater flexibility in branching, even when the final release type
(minor or patch) is unknown.